Reviving pycuda stream engine part 1
